// packages/webdev-api-contracts/src/supabase.schema.ts
export const publicTokenRowSchema = z.object({
  claimed_at: z.string().nullable(),
  claimed_by_id: z.string().nullable(),
  code: z.string(),
  created_at: z.string(),
  creator_id: z.string(),
  id: z.string(),
  is_claimed: z.boolean(),
  is_voided: z.boolean(),
  template_id: z.string(),
});

export const publicTokenInsertSchema = z.object({
  claimed_at: z.string().optional().nullable(),
  claimed_by_id: z.string().optional().nullable(),
  code: z.string(),
  created_at: z.string().optional(),
  creator_id: z.string(),
  id: z.string().optional(),
  is_claimed: z.boolean().optional(),
  is_voided: z.boolean().optional(),
  template_id: z.string(),
});